Solved issues
